docs(Table): dedupe code
