+++ packages/core/src/Table/stories/TableSamples/Main.tsx
@@ -1,0 +1,45 @@
+import {
+  HvTable,
+  HvTableBody,
+  HvTableCell,
+  HvTableContainer,
+  HvTableHead,
+  HvTableHeader,
+  HvTableProps,
+  HvTableRow,
+} from "@hitachivantara/uikit-react-core";
+
+import { getColumns, makeData } from "../storiesUtils";
+
+// props is only used for storybook purposes
+export const Main = (props: Omit<HvTableProps, "children">) => {
+  const columns = getColumns();
+  const data = makeData(6);
+
+  return (
+    <HvTableContainer>
+      <HvTable {...props}>
+        <HvTableHead>
+          <HvTableRow>
+            {columns.map((el) => (
+              <HvTableHeader key={el.Header}>{el.Header}</HvTableHeader>
+            ))}
+          </HvTableRow>
+        </HvTableHead>
+        <HvTableBody>
+          {data.map((el) => (
+            <HvTableRow key={el.id} hover striped>
+              <HvTableCell>{el.name}</HvTableCell>
+              <HvTableCell>{el.createdDate}</HvTableCell>
+              <HvTableCell>{el.eventType}</HvTableCell>
+              <HvTableCell>{el.status}</HvTableCell>
+              <HvTableCell>{el.riskScore}</HvTableCell>
+              <HvTableCell>{el.severity}</HvTableCell>
+              <HvTableCell>{el.priority}</HvTableCell>
+            </HvTableRow>
+          ))}
+        </HvTableBody>
+      </HvTable>
+    </HvTableContainer>
+  );
+};
